pause: add option for multiline input
I came across this issue earlier when trying to store an SSH key using
pause to read a multiline string.

To maintain compatibility, the 'multiline' flag defaults to 'no'.

When setting `multiline: yes`, the prompt will also indicate that
multiple input lines are expected and that the input is terminated by
inserting two (consecutive) newlines.